wip celery tracker refactor
